requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.200
opencv-python>=4.8.0

# Object Tracking
filterpy>=1.4.5
scikit-image>=0.21.0
scipy>=1.11.0

# API and Web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0
aiofiles>=23.2.1

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
pyyaml>=6.0
tqdm>=4.66.0

# Evaluation and Metrics
scikit-learn>=1.3.0
matplotlib>=3.7.0

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
